+++ psRadiantOne/Public/Schemas/Get-R1SchemaTableField.ps1
@@ -1,0 +1,44 @@
+# .ExternalHelp psRadiantOne-help.xml
+function Get-R1SchemaTableField {
+	[CmdletBinding()]
+	[OutputType('psRadiantOne.Field')]
+	param(
+		[parameter(
+			Mandatory = $true,
+			ValueFromPipelineByPropertyName = $true
+		)]
+		[ValidateNotNullOrEmpty()]
+		[Alias('name')]
+		[string]$schemaName,
+
+		[parameter(
+			Mandatory = $true,
+			ValueFromPipelineByPropertyName = $true
+		)]
+		[ValidateNotNullOrEmpty()]
+		[string]$tableName
+	)
+
+	Begin {
+
+		Assert-R1Session -RequireToken
+
+	}#begin
+
+	Process {
+
+		$URI = Resolve-R1ServiceUrl -Service Catalog -Path "schemas/$($schemaName | Get-EscapedString)/tables/$($tableName | Get-EscapedString)/fields"
+
+		$Result = Invoke-R1RestMethod -Uri $URI -Method GET
+
+		if ($null -ne $Result) {
+
+			$Result | Add-CustomType -Type psRadiantOne.Field
+
+		}
+
+	}#process
+
+	End { }#end
+
+}